feat(level-editor): implement UI components for power-ups and waves management
