+++ client/gsm/states/scenes/LevelEditor/UI/CMakeLists.txt
@@ -1,0 +1,8 @@
+target_sources(${CLIENT_BIN_NAME} PRIVATE
+    UI/LevelEditorUI.cpp
+    UI/LevelEditorUIPanel.cpp
+    UI/LevelEditorUIObstacles.cpp
+    UI/LevelEditorUIPowerUps.cpp
+    UI/LevelEditorUIWaves.cpp
+    UI/LevelEditorUIEnemies.cpp
+)
